// fe/src/lib/api/generated/model/studentProfilingReportResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { InventoryCoverage } from './inventoryCoverage';
import type { Methodology } from './methodology';
import type { ProgramColumn } from './programColumn';
import type { ReportContext } from './reportContext';
import type { StudentProfilingSections } from './studentProfilingSections';

export interface StudentProfilingReportResponse {
  inventory_coverage: InventoryCoverage;
  methodology: Methodology;
  program_columns: ProgramColumn[];
  report_context: ReportContext;
  sections: StudentProfilingSections;
}